+++ src/components/Main/PostItem.tsx
@@ -13,6 +13,8 @@ const PostItemWrapper = styled(Link)`
   box-shadow: 0 0 8px rgba(0, 0, 0, 0.15);
   transition: 0.3s box-shadow;
   cursor: pointer;
+  background-color: #eaf2ff;
+  color: #121212;
 
   &:hover {
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
