+++ src/pages/index.tsx
@@ -1,13 +1,14 @@
-import React, { FunctionComponent, useMemo, useEffect, useState } from 'react'
+import React, { FunctionComponent } from 'react'
+import { graphql } from 'gatsby'
+import useThemeToggle from '../hooks/useThemeToggle'
 import CategoryList from 'components/Main/CategoryList'
 import Introduction from 'components/Main/Introduction'
 import PostList from 'components/Main/PostList'
-import { graphql } from 'gatsby'
 import { PostListItemType } from 'types/PostItem.types'
 import { IGatsbyImageData } from 'gatsby-plugin-image'
 import queryString, { ParsedQuery } from 'query-string'
 import Template from 'components/Common/Template'
-import ThemeToggle from '../hooks/themeToggle'
+import ThemeToggle from 'hooks/themeToggle'
 
 import '../styles/lightMode.css'
 import '../styles/darkMode.css'
@@ -62,52 +63,33 @@ const IndexPage: FunctionComponent<IndexPageProps> = function ({
     },
   },
 }) {
-  const [theme, setTheme] = useState<'light' | 'dark'>('light')
+  const { theme, toggleTheme } = useThemeToggle()
 
-  const toggleTheme = () => {
-    const newTheme = theme === 'light' ? 'dark' : 'light'
-    setTheme(newTheme)
-    localStorage.setItem('theme', newTheme)
-  }
-
-  useEffect(() => {
-    const savedTheme = localStorage.getItem('theme') as 'light' | 'dark'
-    if (savedTheme) {
-      setTheme(savedTheme)
-    }
-
-    document.body.className = `theme-${theme}`
-  }, [theme])
-
-  const parsed: ParsedQuery<string> = queryString.parse(search)
+  const parsed: ParsedQuery = queryString.parse(search)
   const selectedCategory: string =
     typeof parsed.category !== 'string' || !parsed.category
       ? 'All'
       : parsed.category
 
-  const categoryList = useMemo(
-    () =>
-      edges.reduce(
-        (
-          list: CategoryListProps['categoryList'],
-          {
-            node: {
-              frontmatter: { categories },
-            },
-          }: PostType,
-        ) => {
-          categories.forEach(category => {
-            if (list[category] === undefined) list[category] = 1
-            else list[category]++
-          })
-
-          list['All']++
-
-          return list
+  const categoryList = edges.reduce(
+    (
+      list: CategoryListProps['categoryList'],
+      {
+        node: {
+          frontmatter: { categories },
         },
-        { All: 0 },
-      ),
-    [],
+      }: PostType,
+    ) => {
+      categories.forEach(category => {
+        if (list[category] === undefined) list[category] = 1
+        else list[category]++
+      })
+
+      list['All']++
+
+      return list
+    },
+    { All: 0 },
   )
 
   return (
